Fix bailout

+++ fluid/tools_fastapi/service.py
@@ -4,6 +4,7 @@ from typing import Any, Self
 
 from fastapi import FastAPI
 
+from fluid import settings
 from fluid.utils import log
 from fluid.utils.worker import Worker, Workers
 
@@ -23,8 +24,9 @@ class FastapiAppWorkers(Workers):
         return workers
 
     def bail_out(self, reason: str, code: int = 1) -> None:
-        logger.warning("shutting down due to %s", reason)
-        os.kill(os.getpid(), signal.SIGTERM)
+        if settings.ENV != "test":
+            logger.warning("shutting down due to %s", reason)
+            os.kill(os.getpid(), signal.SIGTERM)
 
     def get_active_worker(self, *, worker_name: str) -> Worker | None:
         worker = self._workers.get_worker_by_name(worker_name)
